refactor: pick theme refactor from GH-6009
