+++ packages/mask/src/extension/options-page/DashboardDialogs/Base.tsx
@@ -10,9 +10,13 @@ import {
     IconButtonProps,
 } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
-import { ThemeProvider } from '@mui/material/styles'
+import { Theme, ThemeOptions, ThemeProvider } from '@mui/material/styles'
 import CloseIcon from '@mui/icons-material/Close'
-import { extendsTheme, useClassicMaskFullPageTheme, useMatchXS } from '../../../utils'
+import { useMatchXS } from '../../../utils'
+import { useClassicMaskFullPageTheme } from '../../../utils/theme/useClassicMaskFullPageTheme'
+import { appearanceSettings, languageSettings } from '../../../settings/settings'
+import { useValueRef } from '@masknet/shared-base-ui'
+import { cloneDeep, merge } from 'lodash-unified'
 
 const useStyles = makeStyles()((theme) => ({
     root: {
@@ -101,7 +105,7 @@ export function useModal<DialogProps extends object, AdditionalPropsAppendByDisp
         onClose,
     }
     // Restore old theme
-    const theme = useClassicMaskFullPageTheme()
+    const theme = useClassicMaskFullPageTheme(useValueRef(appearanceSettings), useValueRef(languageSettings))
     const renderedComponent =
         state === DialogState.Destroyed ? null : (
             <ThemeProvider theme={theme}>
@@ -160,6 +164,9 @@ const useDashboardDialogWrapperStyles = makeStyles<DashboardDialogWrapperProps>(
     },
 }))
 
+function extendsTheme(extend: (theme: Theme) => ThemeOptions) {
+    return (theme: Theme) => merge(cloneDeep(theme), extend(theme))
+}
 const dialogTheme = extendsTheme((theme) => ({
     components: {
         MuiOutlinedInput: {
