+++ packages/mask/src/extension/popups/UI.tsx
@@ -2,16 +2,18 @@ import { lazy } from 'react'
 import { Navigate, Route, Routes, HashRouter } from 'react-router-dom'
 import { createInjectHooksRenderer, useActivatedPluginsDashboard } from '@masknet/plugin-infra'
 import { PopupRoutes } from '@masknet/shared-base'
-import { useClassicMaskFullPageTheme } from '../../utils'
+import { useClassicMaskFullPageTheme } from '../../utils/theme/useClassicMaskFullPageTheme'
 import '../../social-network-adaptor/browser-action'
 import { Web3Provider } from '@masknet/web3-shared-evm'
 import { PopupWeb3Context } from '../../web3/context'
 import { PopupFrame } from './components/PopupFrame'
 import { Appearance } from '@masknet/theme'
 import { MaskUIRoot } from '../../UIRoot'
+import { useValueRef } from '@masknet/shared-base-ui'
+import { languageSettings } from '../../settings/settings'
 
 function useAlwaysLightTheme() {
-    return useClassicMaskFullPageTheme({ forcePalette: Appearance.light })
+    return useClassicMaskFullPageTheme(Appearance.light, useValueRef(languageSettings))
 }
 const Wallet = lazy(() => import('./pages/Wallet'))
 const Personas = lazy(() => import('./pages/Personas'))
