+++ packages/mask/src/extension/popups/pages/Swap/index.tsx
@@ -9,7 +9,8 @@ import { WalletStateBarUI } from '../../components/WalletStateBar'
 import { SwapBox } from './SwapBox'
 import { SwapWeb3Context } from '../../../../web3/context'
 import { PopupRoutes } from '@masknet/shared-base'
-import { useI18N, usePopupsMaskFullPageTheme } from '../../../../utils'
+import { useI18N } from '../../../../utils'
+import { useSwapPageTheme } from '../../../../utils/theme/useSwapPageTheme'
 import { NetworkPluginID, useReverseAddress } from '@masknet/plugin-infra'
 import { TargetChainIdContext } from '../../../../plugins/Trader/trader/useTargetChainIdContext'
 import { AllProviderTradeContext } from '../../../../plugins/Trader/trader/useAllProviderTradeContext'
@@ -68,7 +69,7 @@ export default function SwapPage() {
     const { t } = useI18N()
     const { classes } = useStyles()
     const chainId = useChainId()
-    const theme = usePopupsMaskFullPageTheme()
+    const theme = useSwapPageTheme()
     const { value: pendingTransactions = [] } = useRecentTransactions({
         status: TransactionStatusType.NOT_DEPEND,
     })
