+++ packages/mask/src/social-network-adaptor/minds.com/customization/custom.ts
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 import { useValueRef } from '@masknet/shared-base-ui'
 import { SubscriptionFromValueRef } from '@masknet/shared-base'
 import type { SocialNetworkUI } from '../../../social-network'
-import { fromRGB, getBackgroundColor, getForegroundColor, shade, toRGB } from '../../../utils/theme-tools'
+import { fromRGB, getBackgroundColor, getForegroundColor, shade, toRGB } from '../../../utils/theme'
 import { themeListItemSelector } from '../utils/selector'
 
 // TODO: get this from DOM. But currently Minds has a single primary color
