+++ packages/mask/src/utils/index.ts
@@ -11,7 +11,6 @@ export * from './i18n-next-ui'
 export * from './messages'
 export * from './permissions'
 export * from './createNormalReactRoot'
-export * from './theme-tools'
 export * from './theme'
 export * from './utils'
 export * from './watcher'
