+++ packages/mask/src/utils/theme/index.ts
@@ -1,0 +1,9 @@
+import './theme-global.d'
+
+// Do not export more files. Those files are deprecated.
+export * from './color-tools'
+export * from './MaskTheme'
+export * from './useColorStyles'
+export * from './useErrorStyles'
+export * from './useThemeLanguage'
+export * from './useClassicMaskSNSTheme'
